minor
solve a warning for casting int32_t int64_t

+++ src/simbuffer.h
@@ -22,7 +22,7 @@ class SimBuffer
 {
 public :
 	static SimBufferPtr Create();
-	int64_t GetNumParticles() { return m_positions.size(); }
+	int32_t GetNumParticles() { return m_positions.size(); }
 
 private :
     SimBuffer() {};
